refactor: move tests to top level
