+++ scripts/integration.sh
@@ -2,9 +2,9 @@
 set -ex
 
 # always run integration tests
-pip install -r jinahub/indexers/tests/requirements.txt
+pip install -r tests/requirements.txt
 export PYTHONPATH=.
-pytest -s -v jinahub/indexers/tests/integration
+pytest -s -v tests/integration
 local_exit_code=$?
 if [[ ! $local_exit_code == 0 ]]; then
   EXIT_CODE=$local_exit_code
