+++ tests/integration/psql_dump_reload/docker-compose.yml
