+++ tests/integration/psql_dump_reload/indexer_dbms.yml
