+++ tests/integration/psql_dump_reload/indexer_query.yml
